projects/jiameng-dianwei/server/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
openai==1.51.0
python-dotenv==1.0.1
apscheduler==3.10.4
alibabacloud-dysmsapi20170525>=2.0.0
pycryptodome>=3.20.0
